fix: DependencyInjectionContainer에서 CafeAppContainer로 클래스명 변경
- 변경 이유는 Cafe 애플리케이션과 관련된 객체를 관리하는 용도이기 때문에 변경하였습니다.
